requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
beautifulsoup4==4.12.3
boto3==1.35.16
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
greenlet==3.1.1
idna==3.10
jmespath==1.0.1
kafka-python==2.0.2
minio==7.2.9
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.21.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
apache-airflow-providers-apache-spark
paho-mqtt